# back/central/shared/docs/auth/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio REST para la gestión de reservas multi-restaurante.",
        "title": "Restaurant Reservation API",
        "termsOfService": "https://ejemplo.com/terminos",
        "contact": {
            "name": "Equipo de Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3050",
    "basePath": "/api/v1",
    "paths": {
...
        "/businesses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene una lista paginada de todos los negocios del sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "Obtener lista de negocios",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página (por defecto 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Elementos por página (por defecto 10)",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por nombre de negocio",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por tipo de negocio",
                        "name": "business_type_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar por estado activo/inactivo",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Negocios obtenidos exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo negocio en el sistema",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "Crear un nuevo negocio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre del negocio",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Código del negocio",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del tipo de negocio",
                        "name": "business_type_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Zona horaria",
                        "name": "timezone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Dirección",
                        "name": "address",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Descripción",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Logo del negocio",
                        "name": "logo_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Color primario",
                        "name": "primary_color",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Color secundario",
                        "name": "secondary_color",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Dominio personalizado",
                        "name": "custom_domain",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Habilitar delivery",
                        "name": "enable_delivery",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Habilitar pickup",
                        "name": "enable_pickup",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Habilitar reservas",
                        "name": "enable_reservations",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Negocio creado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/businesses/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene un negocio específico por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "Obtener negocio por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio obtenido exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.GetBusinessByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un negocio existente",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "Actualizar negocio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nombre del negocio",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Código del negocio",
                        "name": "code",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "ID del tipo de negocio",
                        "name": "business_type_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Zona horaria",
                        "name": "timezone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Dirección",
                        "name": "address",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Descripción",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Logo del negocio",
                        "name": "logo_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Color primario",
                        "name": "primary_color",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Color secundario",
                        "name": "secondary_color",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Dominio personalizado",
                        "name": "custom_domain",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "¿Activo?",
                        "name": "is_active",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Habilitar delivery",
                        "name": "enable_delivery",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Habilitar pickup",
                        "name": "enable_pickup",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Habilitar reservas",
                        "name": "enable_reservations",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Negocio actualizado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina un negocio del sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "Eliminar negocio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Negocio eliminado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.BusinessDetailResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "business_type": {
                    "$ref": "#/definitions/response.BusinessTypeDetailResponse"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "custom_domain": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable_delivery": {
                    "type": "boolean"
                },
                "enable_pickup": {
                    "type": "boolean"
                },
                "enable_reservations": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "navbar_image_url": {
                    "type": "string"
                },
                "primary_color": {
                    "type": "string"
                },
                "quaternary_color": {
                    "type": "string"
                },
                "secondary_color": {
                    "type": "string"
                },
                "tertiary_color": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.BusinessTypeDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.GetBusinessByIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.BusinessDetailResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de business JWT con el prefijo **Bearer** (para todos los endpoints)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BusinessTokenAuth": {
            "description": "Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}